+++ src/styles/theme.tsx
@@ -1,7 +1,9 @@
 export const Theme = {
-  complement: '#B0791E',
-  gray: '#E7E7E7',
-  primary: '#254463',
-  secondary: '#DFEFFF',
-  tertiary: '#95B4D3',
+  borderRadius: '5px',
+  colorComplement: '#B0791E',
+  colorGray: '#E7E7E7',
+  colorPrimary: '#254463',
+  colorSecondary: '#DFEFFF',
+  colorTertiary: '#95B4D3',
+  shadow: '0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)',
 };